# pkg/swagger/swagger/swagger.yml
swagger: "2.0"
info:
  title: Hello
  version: 0.0.1
definitions:
  error:
    type: object
    required:
    - message
    properties:
      code:
        type: integer
        format: int64
      message:
        type: string
  user:
    type: object
    title: User represents the user for this application
    description: "A user is the security principal for this application.\nIt's also used as one of the main axes for reporting.\n\nA user can have friends with whom they can share what they like."
    required:
    - id
    properties:
      id:
        description: the id for this user
        type: integer
        format: int64
        minimum: 1
      name:
        description: the name for this user
        type: string
        minLength: 3
paths:
  /hostname:
    get:
      operationId: getHostname
      produces:
      - text/plain
      responses:
        200:
          description: returns the hostname of the machine
          schema:
            description: the hostname of the machine
            type: string
        default:
          description: error
          schema:
            $ref: "#/definitions/error"
  /users:
    get:
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      produces:
        - application/json
      responses:
        '200':    # status code
          description: A JSON array of user names
          schema:
            type: array
            items:
              type: object
              $ref: "#/definitions/user"
  /user/{userId}:
    get:
      summary: Returns a user by ID.
      parameters:
      - name: userId
        in: path
        required: true
        description: The ID of the user to return.
        type: integer
        format: int64
        minimum: 1
      produces:
      - application/json
      responses:
        '200':
          description: A user object.
          schema:
            type: object
            $ref: "#/definitions/user"
        '400':
          description: The specified user ID is invalid (not a number).
        '404':
          description: A user with the specified ID was not found.
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/error"